+++ CHANGES.rst
@@ -4,6 +4,14 @@ Change Log
 Unreleased
 ----------
 
+**Changed**
+
+- Updated black config parsing.
+
+**Fixed**
+
+- Fixed import error when parsing black config.
+
 1.1.0 (2021/02/18)
 ------------------